Switch to storing only window in the state file
This commit changes the state file from storing a desktop and a window in the state file to storing only the window. There was no reason to store the desktop as it may have changed if the window was moved to another desktop since it was last seen, and the desktop for the window can be queried at the time to activate it by getting the _NET_WM_DESKTOP property.

This changes the state file format on disk as it now stores only 2 32-bit integers instead of 4. A future commit will add a version header to the state file.